+++ R/extractors.R
@@ -19,12 +19,14 @@ package_by.cranly_network <- function(x, author = NULL, exact = FALSE) {
     }
     inds <- sapply(x$nodes$author, function(z) any(grepl(str, z, ignore.case = !exact)))
     out <- unique(unlist(x$nodes[inds, "package"]))
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
+
 }
 
 #' @rdname package_by
@@ -49,12 +51,14 @@ package_with.cranly_network <- function(x, name = NULL, exact = FALSE) {
     package <- unlist(x$nodes$package)
     inds <- grep(str, package, ignore.case = !exact, perl = TRUE)
     out <- unique(package[inds])
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
+
 }
 
 #' @rdname package_by
@@ -77,11 +81,12 @@ author_of.cranly_network <- function(x, package = NULL, exact = FALSE) {
     }
     inds <- grep(str, x$nodes$package, ignore.case = !exact, perl = TRUE)
     out <- unique(unlist(x$nodes[inds, "author"]))
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
 }
 
@@ -105,13 +110,45 @@ author_with.cranly_network <- function(x, name = NULL, exact = FALSE) {
     authors <- unlist(x$nodes$author)
     inds <- grep(str, authors, ignore.case = !exact)
     out <- unique(authors[inds])
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
 }
+
+#' @rdname package_by
+#' @export
+suggests <- function(x, package = NULL, exact = FALSE) {
+    if (is.null(package)) {
+        return(NULL) # return(unlist(x$nodes$Package))
+    }
+    if (any(is.infinite(package))) {
+        return(unique(unlist(x$nodes$suggests)))
+    }
+    perspective <- attr(x, "perspective")
+    package <- gsub("\\.", "\\\\.", package)
+    if (exact) {
+        str <- paste(package, collapse = "$(?!\\.)|^")
+        str <- paste0("^", str, "$(?!\\.)")
+    }
+    else {
+        str <- paste(package, collapse = "|")
+    }
+    inds <- grep(str, x$nodes$package, ignore.case = !exact, perl = TRUE)
+    ## inds <- sapply(x$nodes$Package, function(x) any(grepl(str, x)))
+    out <- unique(unlist(x$nodes[inds, "suggests"]))
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
+    }
+    else {
+        return(out)
+    }
+}
+
 
 #' @rdname package_by
 #' @export
@@ -134,11 +171,12 @@ imports <- function(x, package = NULL, exact = FALSE) {
     inds <- grep(str, x$nodes$package, ignore.case = !exact, perl = TRUE)
     ## inds <- sapply(x$nodes$Package, function(x) any(grepl(str, x)))
     out <- unique(unlist(x$nodes[inds, "imports"]))
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
 }
 
@@ -163,11 +201,12 @@ depends <- function(x, package = NULL, exact = FALSE) {
     inds <- grep(str, x$nodes$package, ignore.case = !exact, perl = TRUE)
     ## inds <- sapply(x$nodes$Package, function(x) any(grepl(str, x)))
     out <- unique(unlist(x$nodes[inds, "depends"]))
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
 }
 
@@ -193,11 +232,12 @@ linking_to <- function(x, package = NULL, exact = FALSE) {
     inds <- grep(str, x$nodes$package, ignore.case = !exact, perl = TRUE)
     ## inds <- sapply(x$nodes$Package, function(x) any(grepl(str, x)))
     out <- unique(unlist(x$nodes[inds, "linkingto"]))
-    if (length(out)) {
-        out
+
+    if (all(is.na(out)) | !length(out)) {
+        return(NULL)
     }
     else {
-        NULL
+        return(out)
     }
 }